Move AttributeOptionGroup dashboard view tests to different module
